+++ src/Resources/app/storefront/src/register.js
@@ -1,12 +1,11 @@
 import MollieCreditCardComponents from './mollie-payments/plugins/creditcard-components.plugin';
 import MollieCreditCardComponentsSw64 from './mollie-payments/plugins/creditcard-components-sw64.plugin';
-import MollieApplePayDirect from './mollie-payments/plugins/apple-pay-direct.plugin';
+import MollieApplePayDirect from './mollie-payments/plugins/express/apple-pay-direct.plugin';
 import MollieApplePayPaymentMethod from './mollie-payments/plugins/apple-pay-payment-method.plugin';
 import MollieCreditCardMandateManage from './mollie-payments/plugins/creditcard-mandate-manage.plugin';
 import MolliePosTerminalPlugin from './mollie-payments/plugins/pos-terminal.plugin';
-import PayPalExpressPlugin from './mollie-payments/plugins/paypal-express.plugin';
+import PayPalExpressPlugin from './mollie-payments/plugins/express/paypal-express.plugin';
 import MollieBancomatPlugin from './mollie-payments/plugins/bancomat-plugin';
-import { MollieExpressActions } from './mollie-payments/plugins/mollie-express-actions.plugin';
 
 export default class MollieRegistration {
     /**
@@ -18,7 +17,6 @@ export default class MollieRegistration {
         // global plugins
         // -----------------------------------------------------------------------------
         // hide apple pay direct buttons across the whole shop, if not available
-        pluginManager.register('MollieExpressActions', MollieExpressActions);
         pluginManager.register('MollieApplePayDirect', MollieApplePayDirect);
 
         // fix quantity select on PDP Page