+++ tests/Cypress/cypress/e2e/storefront/payment-methods/paypal-express.cy.js
@@ -9,6 +9,7 @@ import CheckoutAction from "Actions/storefront/checkout/CheckoutAction";
 import CartRepository from "Repositories/storefront/checkout/CartRepository";
 import ListingRepository from "Repositories/storefront/products/ListingRepository";
 import RegisterRepository from "Repositories/storefront/checkout/RegisterRepository";
+import ShopConfiguration from "../../../support/models/ShopConfiguration";
 
 
 const devices = new Devices();
@@ -25,9 +26,19 @@ const repoCart = new CartRepository();
 const registerRepo = new RegisterRepository();
 
 
-function beforeEach() {
+function beforeEachNoPrivacy() {
     cy.wrap(null).then(() => {
         devices.setDevice(devices.getFirstDevice());
+        configAction.setupShop(true, false, false, new ShopConfiguration());
+    });
+}
+
+function beforeEachPrivacy() {
+    cy.wrap(null).then(() => {
+        devices.setDevice(devices.getFirstDevice());
+        const config = new ShopConfiguration();
+        config.setDataPrivacy(true);
+        configAction.setupShop(true, false, false, config);
     });
 }
 
@@ -36,26 +47,28 @@ describe('Paypal Express - UI Tests', () => {
 
     describe('PDP', () => {
 
-        it('Paypal Express button is visible @core', () => {
+        it('C4247553: Paypal Express button is visible @core', () => {
 
-            beforeEach();
+            beforeEachNoPrivacy();
 
-            configAction.setupPlugin(false, false, false, false, []);
+            cy.wrap(null).then(() => {
+                configAction.setupPlugin(false, false, false, false, []);
+            });
 
             cy.visit('/');
             topMenu.clickOnSecondCategory();
             listing.clickOnFirstProduct();
 
-            const button = repoPDP.getPayPalExpressButton();
-            button.should('be.visible');
-
+            repoPDP.getPayPalExpressButton().should('be.visible');
         })
 
-        it('Paypal Express button is hidden because of restriction @core', () => {
+        it('C4247549: Paypal Express button is hidden because of restriction @core', () => {
 
-            beforeEach();
+            beforeEachNoPrivacy();
 
-            configAction.setupPlugin(false, false, false, false, ['pdp']);
+            cy.wrap(null).then(() => {
+                configAction.setupPlugin(false, false, false, false, ['pdp']);
+            });
 
             cy.visit('/');
             topMenu.clickOnSecondCategory();
@@ -64,15 +77,34 @@ describe('Paypal Express - UI Tests', () => {
             repoPDP.getPayPalExpressButton().should('not.exist');
         })
 
+        it('C4247548: PayPal Express requires data protection to be accepted if enabled @core', () => {
+
+            beforeEachPrivacy();
+
+            cy.wrap(null).then(() => {
+                configAction.setupPlugin(false, false, false, false, []);
+            });
+
+            cy.visit('/');
+            topMenu.clickOnSecondCategory();
+            listing.clickOnFirstProduct();
+
+            // click and make sure data privacy is validated
+            repoPDP.getPayPalExpressButton().click();
+            repoPDP.getDataPrivacyCheckbox().should('have.class', 'is-invalid');
+        })
+
     })
 
     describe('Listing', () => {
 
-        it('Paypal Express button is visible @core', () => {
+        it('C4247554: Paypal Express button is visible @core', () => {
 
-            beforeEach();
+            beforeEachNoPrivacy();
 
-            configAction.setupPlugin(false, false, false, false, []);
+            cy.wrap(null).then(() => {
+                configAction.setupPlugin(false, false, false, false, []);
+            });
 
             cy.visit('/');
             topMenu.clickOnSecondCategory();
@@ -82,27 +114,47 @@ describe('Paypal Express - UI Tests', () => {
 
         })
 
-        it('Paypal Express button is hidden because of restriction @core', () => {
+        it('C4247550: Paypal Express button is hidden because of restriction @core', () => {
 
-            beforeEach();
+            beforeEachNoPrivacy();
 
-            configAction.setupPlugin(false, false, false, false, ['plp']);
+            cy.wrap(null).then(() => {
+                configAction.setupPlugin(false, false, false, false, ['plp']);
+            });
 
             cy.visit('/');
             topMenu.clickOnSecondCategory();
 
             repoListing.getPayPalExpressButton().should('not.exist');
         })
-    })
 
+        it('C4247547: PayPal Express requires data protection to be accepted if enabled @core', () => {
+
+            beforeEachPrivacy();
+
+            cy.wrap(null).then(() => {
+                configAction.setupPlugin(false, false, false, false, []);
+            });
+
+            cy.visit('/');
+            topMenu.clickOnSecondCategory();
+
+            // click and make sure data privacy is validated
+            repoListing.getPayPalExpressButton().first().click();
+            repoListing.getDataPrivacyCheckbox().first().should('have.class', 'is-invalid');
+        })
+
+    })
 
     describe('Offcanvas', () => {
 
-        it('Paypal Express button is visible @core', () => {
+        it('C4247556: Paypal Express button is visible @core', () => {
 
-            beforeEach();
+            beforeEachNoPrivacy();
 
-            configAction.setupPlugin(false, false, false, false, []);
+            cy.wrap(null).then(() => {
+                configAction.setupPlugin(false, false, false, false, []);
+            });
 
             cy.visit('/');
             topMenu.clickOnSecondCategory();
@@ -113,11 +165,13 @@ describe('Paypal Express - UI Tests', () => {
             button.should('be.visible');
         })
 
-        it('Paypal Express button is hidden because of restriction @core', () => {
+        it('C4247551: Paypal Express button is hidden because of restriction @core', () => {
 
-            beforeEach();
+            beforeEachNoPrivacy();
 
-            configAction.setupPlugin(false, false, false, false, ['offcanvas'])
+            cy.wrap(null).then(() => {
+                configAction.setupPlugin(false, false, false, false, ['offcanvas'])
+            });
 
             cy.visit('/');
             topMenu.clickOnSecondCategory();
@@ -127,15 +181,35 @@ describe('Paypal Express - UI Tests', () => {
             repoOffcanvas.getPayPalExpressButton().should('not.exist');
         })
 
+        it('C4247546: PayPal Express requires data protection to be accepted if enabled @core', () => {
+
+            beforeEachPrivacy();
+
+            cy.wrap(null).then(() => {
+                configAction.setupPlugin(false, false, false, false, []);
+            });
+
+            cy.visit('/');
+            topMenu.clickOnSecondCategory();
+            listing.clickOnFirstProduct();
+            pdp.addToCart(1);
+
+            // click and make sure data privacy is validated
+            repoOffcanvas.getPayPalExpressButton().click();
+            repoOffcanvas.getDataPrivacyCheckbox().should('have.class', 'is-invalid');
+        })
+
     })
 
     describe('Cart', () => {
 
-        it('Paypal Express button is visible @core', () => {
+        it('C4247555: Paypal Express button is visible @core', () => {
 
-            beforeEach();
+            beforeEachNoPrivacy();
 
-            configAction.setupPlugin(false, false, false, false, []);
+            cy.wrap(null).then(() => {
+                configAction.setupPlugin(false, false, false, false, []);
+            });
 
             cy.visit('/');
             topMenu.clickOnSecondCategory();
@@ -149,11 +223,13 @@ describe('Paypal Express - UI Tests', () => {
 
         })
 
-        it('Paypal Express button is hidden because of restriction @core', () => {
+        it('C4247552: Paypal Express button is hidden because of restriction @core', () => {
 
-            beforeEach();
+            beforeEachNoPrivacy();
 
-            configAction.setupPlugin(false, false, false, false, ['cart']);
+            cy.wrap(null).then(() => {
+                configAction.setupPlugin(false, false, false, false, ['cart']);
+            });
 
             cy.visit('/');
             topMenu.clickOnSecondCategory();
@@ -165,15 +241,37 @@ describe('Paypal Express - UI Tests', () => {
             repoCart.getPayPalExpressButton().should('not.exist');
         })
 
+        it('C4247545: PayPal Express requires data protection to be accepted if enabled @core', () => {
+
+            beforeEachPrivacy();
+
+            cy.wrap(null).then(() => {
+                configAction.setupPlugin(false, false, false, false, []);
+            });
+
+            cy.visit('/');
+            topMenu.clickOnSecondCategory();
+            listing.clickOnFirstProduct();
+            pdp.addToCart(1);
+
+            checkout.goToCartInOffCanvas();
+
+            // click and make sure data privacy is validated
+            repoCart.getPayPalExpressButton().click();
+            repoCart.getDataPrivacyCheckbox().should('have.class', 'is-invalid');
+        })
+
     })
 
     describe('Register Page', () => {
 
-        it('Paypal Express button is visible @core', () => {
+        it('C4247557: Paypal Express button is visible @core', () => {
 
-            beforeEach();
+            beforeEachNoPrivacy();
 
-            configAction.setupPlugin(false, false, false, false, []);
+            cy.wrap(null).then(() => {
+                configAction.setupPlugin(false, false, false, false, []);
+            });
 
             cy.visit('/');
             topMenu.clickOnSecondCategory();
@@ -187,12 +285,14 @@ describe('Paypal Express - UI Tests', () => {
 
         })
 
-        it('Paypal Express button is hidden because of restriction @core', () => {
+        it('C4247558: Paypal Express button is hidden because of restriction @core', () => {
 
-            beforeEach();
+            beforeEachNoPrivacy();
 
-            configAction.setupPlugin(false, false, false, false, ['register']);
-
+            cy.wrap(null).then(() => {
+                configAction.setupPlugin(false, false, false, false, ['register']);
+            });
+            
             cy.visit('/');
             topMenu.clickOnSecondCategory();
             listing.clickOnFirstProduct();